lp-flask-python/requirements.txt
dynaconf~=3.2.6
flask~=3.0.3
python-jose~=3.3.0
Flask-SQLAlchemy~=3.1.1